move object spawning from NetworkClient into NetworkObjectManager
