+++ Assets/Mirror/Runtime/NetworkClient.cs
@@ -157,7 +157,6 @@ namespace Mirror
             {
                 IConnection transportConnection = await transport.ConnectAsync(uri);
 
-                RegisterSpawnPrefabs();
                 InitializeAuthEvents();
 
                 // setup all the handlers
@@ -294,24 +293,12 @@ namespace Mirror
 
         internal void RegisterHostHandlers()
         {
-            Connection.RegisterHandler<ObjectDestroyMessage>(OnHostClientObjectDestroy);
-            Connection.RegisterHandler<ObjectHideMessage>(OnHostClientObjectHide);
             Connection.RegisterHandler<NetworkPongMessage>(msg => { });
-            Connection.RegisterHandler<SpawnMessage>(OnHostClientSpawn);
-            // host mode reuses objects in the server
-            // so we don't need to spawn them
-            Connection.RegisterHandler<UpdateVarsMessage>(msg => { });
-            Connection.RegisterHandler<RpcMessage>(OnRpcMessage);
         }
 
         internal void RegisterMessageHandlers()
         {
-            Connection.RegisterHandler<ObjectDestroyMessage>(OnObjectDestroy);
-            Connection.RegisterHandler<ObjectHideMessage>(OnObjectHide);
             Connection.RegisterHandler<NetworkPongMessage>(Time.OnClientPong);
-            Connection.RegisterHandler<SpawnMessage>(OnSpawn);
-            Connection.RegisterHandler<UpdateVarsMessage>(OnUpdateVarsMessage);
-            Connection.RegisterHandler<RpcMessage>(OnRpcMessage);
         }
 
         /// <summary>
@@ -322,8 +309,6 @@ namespace Mirror
         {
             logger.Log("Shutting down client.");
 
-            ClearSpawners();
-            DestroyAllClientObjects();
             hostServer = null;
 
             connectState = ConnectState.Disconnected;
@@ -338,497 +323,6 @@ namespace Mirror
             {
                 // if no authenticator, consider connection as authenticated
                 Connected.RemoveListener(OnAuthenticated);
-            }
-        }
-
-        static bool ConsiderForSpawning(NetworkIdentity identity)
-        {
-            // not spawned yet, not hidden, etc.?
-            return !identity.gameObject.activeSelf &&
-                   identity.gameObject.hideFlags != HideFlags.NotEditable &&
-                   identity.gameObject.hideFlags != HideFlags.HideAndDontSave &&
-                   identity.sceneId != 0;
-        }
-
-        // this is called from message handler for Owner message
-        internal void InternalAddPlayer(NetworkIdentity identity)
-        {
-            if (Connection != null)
-            {
-                Connection.Identity = identity;
-            }
-            else
-            {
-                logger.LogWarning("No ready connection found for setting player controller during InternalAddPlayer");
-            }
-        }
-
-        /// <summary>
-        /// Call this after loading/unloading a scene in the client after connection to register the spawnable objects
-        /// </summary>
-        public void PrepareToSpawnSceneObjects()
-        {
-            // add all unspawned NetworkIdentities to spawnable objects
-            spawnableObjects.Clear();
-            IEnumerable<NetworkIdentity> sceneObjects =
-                Resources.FindObjectsOfTypeAll<NetworkIdentity>()
-                               .Where(ConsiderForSpawning);
-
-            foreach (NetworkIdentity obj in sceneObjects)
-            {
-                spawnableObjects.Add(obj.sceneId, obj);
-            }
-        }
-
-        #region Spawn Prefabs
-        private void RegisterSpawnPrefabs()
-        {
-            for (int i = 0; i < spawnPrefabs.Count; i++)
-            {
-                GameObject prefab = spawnPrefabs[i];
-                if (prefab != null)
-                {
-                    RegisterPrefab(prefab);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Find the registered prefab for this asset id.
-        /// Useful for debuggers
-        /// </summary>
-        /// <param name="assetId">asset id of the prefab</param>
-        /// <param name="prefab">the prefab gameobject</param>
-        /// <returns>true if prefab was registered</returns>
-        public GameObject GetPrefab(Guid assetId)
-        {
-            if (assetId == Guid.Empty)
-                return null;
-
-            if (prefabs.TryGetValue(assetId, out GameObject prefab))
-            {
-                return prefab;
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// Registers a prefab with the spawning system.
-        /// <para>When a NetworkIdentity object is spawned on a server with NetworkServer.SpawnObject(), and the prefab that the object was created from was registered with RegisterPrefab(), the client will use that prefab to instantiate a corresponding client object with the same netId.</para>
-        /// <para>The NetworkManager has a list of spawnable prefabs, it uses this function to register those prefabs with the ClientScene.</para>
-        /// <para>The set of current spawnable object is available in the ClientScene static member variable ClientScene.prefabs, which is a dictionary of NetworkAssetIds and prefab references.</para>
-        /// </summary>
-        /// <param name="prefab">A Prefab that will be spawned.</param>
-        /// <param name="newAssetId">An assetId to be assigned to this prefab. This allows a dynamically created game object to be registered for an already known asset Id.</param>
-        public void RegisterPrefab(GameObject prefab, Guid newAssetId)
-        {
-            NetworkIdentity identity = prefab.GetComponent<NetworkIdentity>();
-            if (identity)
-            {
-                identity.AssetId = newAssetId;
-
-                if (logger.LogEnabled()) logger.Log("Registering prefab '" + prefab.name + "' as asset:" + identity.AssetId);
-                prefabs[identity.AssetId] = prefab;
-            }
-            else
-            {
-                throw new InvalidOperationException("Could not register '" + prefab.name + "' since it contains no NetworkIdentity component");
-            }
-        }
-
-        /// <summary>
-        /// Registers a prefab with the spawning system.
-        /// <para>When a NetworkIdentity object is spawned on a server with NetworkServer.SpawnObject(), and the prefab that the object was created from was registered with RegisterPrefab(), the client will use that prefab to instantiate a corresponding client object with the same netId.</para>
-        /// <para>The NetworkManager has a list of spawnable prefabs, it uses this function to register those prefabs with the ClientScene.</para>
-        /// <para>The set of current spawnable object is available in the ClientScene static member variable ClientScene.prefabs, which is a dictionary of NetworkAssetIds and prefab references.</para>
-        /// </summary>
-        /// <param name="prefab">A Prefab that will be spawned.</param>
-        public void RegisterPrefab(GameObject prefab)
-        {
-            NetworkIdentity identity = prefab.GetComponent<NetworkIdentity>();
-            if (identity)
-            {
-                if (logger.LogEnabled()) logger.Log("Registering prefab '" + prefab.name + "' as asset:" + identity.AssetId);
-                prefabs[identity.AssetId] = prefab;
-
-                NetworkIdentity[] identities = prefab.GetComponentsInChildren<NetworkIdentity>();
-                if (identities.Length > 1)
-                {
-                    logger.LogWarning("The prefab '" + prefab.name +
-                                     "' has multiple NetworkIdentity components. There can only be one NetworkIdentity on a prefab, and it must be on the root object.");
-                }
-            }
-            else
-            {
-                throw new InvalidOperationException("Could not register '" + prefab.name + "' since it contains no NetworkIdentity component");
-            }
-        }
-
-        /// <summary>
-        /// Registers a prefab with the spawning system.
-        /// <para>When a NetworkIdentity object is spawned on a server with NetworkServer.SpawnObject(), and the prefab that the object was created from was registered with RegisterPrefab(), the client will use that prefab to instantiate a corresponding client object with the same netId.</para>
-        /// <para>The NetworkManager has a list of spawnable prefabs, it uses this function to register those prefabs with the ClientScene.</para>
-        /// <para>The set of current spawnable object is available in the ClientScene static member variable ClientScene.prefabs, which is a dictionary of NetworkAssetIds and prefab references.</para>
-        /// </summary>
-        /// <param name="prefab">A Prefab that will be spawned.</param>
-        /// <param name="spawnHandler">A method to use as a custom spawnhandler on clients.</param>
-        /// <param name="unspawnHandler">A method to use as a custom un-spawnhandler on clients.</param>
-        public void RegisterPrefab(GameObject prefab, SpawnDelegate spawnHandler, UnSpawnDelegate unspawnHandler)
-        {
-            RegisterPrefab(prefab, msg => spawnHandler(msg.position, msg.assetId), unspawnHandler);
-        }
-
-        /// <summary>
-        /// Registers a prefab with the spawning system.
-        /// <para>When a NetworkIdentity object is spawned on a server with NetworkServer.SpawnObject(), and the prefab that the object was created from was registered with RegisterPrefab(), the client will use that prefab to instantiate a corresponding client object with the same netId.</para>
-        /// <para>The NetworkManager has a list of spawnable prefabs, it uses this function to register those prefabs with the ClientScene.</para>
-        /// <para>The set of current spawnable object is available in the ClientScene static member variable ClientScene.prefabs, which is a dictionary of NetworkAssetIds and prefab references.</para>
-        /// </summary>
-        /// <param name="prefab">A Prefab that will be spawned.</param>
-        /// <param name="spawnHandler">A method to use as a custom spawnhandler on clients.</param>
-        /// <param name="unspawnHandler">A method to use as a custom un-spawnhandler on clients.</param>
-        public void RegisterPrefab(GameObject prefab, SpawnHandlerDelegate spawnHandler, UnSpawnDelegate unspawnHandler)
-        {
-            NetworkIdentity identity = prefab.GetComponent<NetworkIdentity>();
-            if (identity == null)
-            {
-                throw new InvalidOperationException("Could not register '" + prefab.name + "' since it contains no NetworkIdentity component");
-            }
-
-            if (identity.AssetId == Guid.Empty)
-            {
-                throw new InvalidOperationException("RegisterPrefab game object " + prefab.name + " has no " + nameof(prefab) + ". Use RegisterSpawnHandler() instead?");
-            }
-
-            if (logger.LogEnabled()) logger.Log("Registering custom prefab '" + prefab.name + "' as asset:" + identity.AssetId + " " + spawnHandler.GetMethodName() + "/" + unspawnHandler.GetMethodName());
-
-            spawnHandlers[identity.AssetId] = spawnHandler;
-            unspawnHandlers[identity.AssetId] = unspawnHandler;
-        }
-
-        /// <summary>
-        /// Removes a registered spawn prefab that was setup with ClientScene.RegisterPrefab.
-        /// </summary>
-        /// <param name="prefab">The prefab to be removed from registration.</param>
-        public void UnregisterPrefab(GameObject prefab)
-        {
-            NetworkIdentity identity = prefab.GetComponent<NetworkIdentity>();
-            if (identity == null)
-            {
-                throw new InvalidOperationException("Could not unregister '" + prefab.name + "' since it contains no NetworkIdentity component");
-            }
-            spawnHandlers.Remove(identity.AssetId);
-            unspawnHandlers.Remove(identity.AssetId);
-        }
-
-        #endregion
-
-        #region Spawn Handler
-
-        /// <summary>
-        /// This is an advanced spawning function that registers a custom assetId with the UNET spawning system.
-        /// <para>This can be used to register custom spawning methods for an assetId - instead of the usual method of registering spawning methods for a prefab. This should be used when no prefab exists for the spawned objects - such as when they are constructed dynamically at runtime from configuration data.</para>
-        /// </summary>
-        /// <param name="assetId">Custom assetId string.</param>
-        /// <param name="spawnHandler">A method to use as a custom spawnhandler on clients.</param>
-        /// <param name="unspawnHandler">A method to use as a custom un-spawnhandler on clients.</param>
-        public void RegisterSpawnHandler(Guid assetId, SpawnDelegate spawnHandler, UnSpawnDelegate unspawnHandler)
-        {
-            RegisterSpawnHandler(assetId, msg => spawnHandler(msg.position, msg.assetId), unspawnHandler);
-        }
-
-        /// <summary>
-        /// This is an advanced spawning function that registers a custom assetId with the UNET spawning system.
-        /// <para>This can be used to register custom spawning methods for an assetId - instead of the usual method of registering spawning methods for a prefab. This should be used when no prefab exists for the spawned objects - such as when they are constructed dynamically at runtime from configuration data.</para>
-        /// </summary>
-        /// <param name="assetId">Custom assetId string.</param>
-        /// <param name="spawnHandler">A method to use as a custom spawnhandler on clients.</param>
-        /// <param name="unspawnHandler">A method to use as a custom un-spawnhandler on clients.</param>
-        public void RegisterSpawnHandler(Guid assetId, SpawnHandlerDelegate spawnHandler, UnSpawnDelegate unspawnHandler)
-        {
-            if (logger.LogEnabled()) logger.Log("RegisterSpawnHandler asset '" + assetId + "' " + spawnHandler.GetMethodName() + "/" + unspawnHandler.GetMethodName());
-
-            spawnHandlers[assetId] = spawnHandler;
-            unspawnHandlers[assetId] = unspawnHandler;
-        }
-
-        /// <summary>
-        /// Removes a registered spawn handler function that was registered with ClientScene.RegisterHandler().
-        /// </summary>
-        /// <param name="assetId">The assetId for the handler to be removed for.</param>
-        public void UnregisterSpawnHandler(Guid assetId)
-        {
-            spawnHandlers.Remove(assetId);
-            unspawnHandlers.Remove(assetId);
-        }
-
-        /// <summary>
-        /// This clears the registered spawn prefabs and spawn handler functions for this client.
-        /// </summary>
-        public void ClearSpawners()
-        {
-            prefabs.Clear();
-            spawnHandlers.Clear();
-            unspawnHandlers.Clear();
-        }
-
-        #endregion
-
-        void UnSpawn(NetworkIdentity identity)
-        {
-            Guid assetId = identity.AssetId;
-
-            identity.StopClient();
-            if (unspawnHandlers.TryGetValue(assetId, out UnSpawnDelegate handler) && handler != null)
-            {
-                handler(identity.gameObject);
-            }
-            else if (identity.sceneId == 0)
-            {
-                Destroy(identity.gameObject);
-            }
-            else
-            {
-                identity.Reset();
-                identity.gameObject.SetActive(false);
-                spawnableObjects[identity.sceneId] = identity;
-            }
-        }
-
-        /// <summary>
-        /// Destroys all networked objects on the client.
-        /// <para>This can be used to clean up when a network connection is closed.</para>
-        /// </summary>
-        public void DestroyAllClientObjects()
-        {
-            foreach (NetworkIdentity identity in Spawned.Values)
-            {
-                if (identity != null && identity.gameObject != null)
-                {
-                    UnSpawn(identity);
-                }
-            }
-            Spawned.Clear();
-        }
-
-        void ApplySpawnPayload(NetworkIdentity identity, SpawnMessage msg)
-        {
-            if (msg.assetId != Guid.Empty)
-                identity.AssetId = msg.assetId;
-
-            if (!identity.gameObject.activeSelf)
-            {
-                identity.gameObject.SetActive(true);
-            }
-
-            // apply local values for VR support
-            identity.transform.localPosition = msg.position;
-            identity.transform.localRotation = msg.rotation;
-            identity.transform.localScale = msg.scale;
-            identity.HasAuthority = msg.isOwner;
-            identity.NetId = msg.netId;
-            identity.Client = this;
-
-            if (msg.isLocalPlayer)
-                InternalAddPlayer(identity);
-
-            // deserialize components if any payload
-            // (Count is 0 if there were no components)
-            if (msg.payload.Count > 0)
-            {
-                using (PooledNetworkReader payloadReader = NetworkReaderPool.GetReader(msg.payload))
-                {
-                    identity.OnDeserializeAllSafely(payloadReader, true);
-                }
-            }
-
-            Spawned[msg.netId] = identity;
-
-            // objects spawned as part of initial state are started on a second pass
-            identity.NotifyAuthority();
-            identity.StartClient();
-            CheckForLocalPlayer(identity);
-        }
-
-        internal void OnSpawn(SpawnMessage msg)
-        {
-            if (msg.assetId == Guid.Empty && msg.sceneId == 0)
-            {
-                throw new InvalidOperationException("OnObjSpawn netId: " + msg.netId + " has invalid asset Id");
-            }
-            if (logger.LogEnabled()) logger.Log($"Client spawn handler instantiating netId={msg.netId} assetID={msg.assetId} sceneId={msg.sceneId} pos={msg.position}");
-
-            // was the object already spawned?
-            NetworkIdentity identity = GetExistingObject(msg.netId);
-
-            if (identity == null)
-            {
-                identity = msg.sceneId == 0 ? SpawnPrefab(msg) : SpawnSceneObject(msg);
-            }
-
-            if (identity == null)
-            {
-                throw new InvalidOperationException($"Could not spawn assetId={msg.assetId} scene={msg.sceneId} netId={msg.netId}");
-            }
-
-            ApplySpawnPayload(identity, msg);
-        }
-
-        NetworkIdentity GetExistingObject(uint netid)
-        {
-            Spawned.TryGetValue(netid, out NetworkIdentity localObject);
-            return localObject;
-        }
-
-        NetworkIdentity SpawnPrefab(SpawnMessage msg)
-        {
-            if (spawnHandlers.TryGetValue(msg.assetId, out SpawnHandlerDelegate handler))
-            {
-                GameObject obj = handler(msg);
-                if (obj == null)
-                {
-                    logger.LogWarning("Client spawn handler for " + msg.assetId + " returned null");
-                    return null;
-                }
-                return obj.GetComponent<NetworkIdentity>();
-            }
-            GameObject prefab = GetPrefab(msg.assetId);
-            if ((object)prefab != null)
-            {
-                GameObject obj = Object.Instantiate(prefab, msg.position, msg.rotation);
-                if (logger.LogEnabled())
-                {
-                    logger.Log("Client spawn handler instantiating [netId:" + msg.netId + " asset ID:" + msg.assetId + " pos:" + msg.position + " rotation: " + msg.rotation + "]");
-                }
-
-                return obj.GetComponent<NetworkIdentity>();
-            }
-            logger.LogError("Failed to spawn server object, did you forget to add it to the NetworkManager? assetId=" + msg.assetId + " netId=" + msg.netId);
-            return null;
-        }
-
-        internal NetworkIdentity SpawnSceneObject(SpawnMessage msg)
-        {
-            NetworkIdentity spawnedId = SpawnSceneObject(msg.sceneId);
-            if (spawnedId == null)
-            {
-                logger.LogError("Spawn scene object not found for " + msg.sceneId.ToString("X") + " SpawnableObjects.Count=" + spawnableObjects.Count);
-
-                // dump the whole spawnable objects dict for easier debugging
-                if (logger.LogEnabled())
-                {
-                    foreach (KeyValuePair<ulong, NetworkIdentity> kvp in spawnableObjects)
-                        logger.Log("Spawnable: SceneId=" + kvp.Key + " name=" + kvp.Value.name);
-                }
-            }
-
-            if (logger.LogEnabled()) logger.Log("Client spawn for [netId:" + msg.netId + "] [sceneId:" + msg.sceneId + "] obj:" + spawnedId);
-            return spawnedId;
-        }
-
-        NetworkIdentity SpawnSceneObject(ulong sceneId)
-        {
-            if (spawnableObjects.TryGetValue(sceneId, out NetworkIdentity identity))
-            {
-                spawnableObjects.Remove(sceneId);
-                return identity;
-            }
-            logger.LogWarning("Could not find scene object with sceneid:" + sceneId.ToString("X"));
-            return null;
-        }
-
-        internal void OnObjectHide(ObjectHideMessage msg)
-        {
-            DestroyObject(msg.netId);
-        }
-
-        internal void OnObjectDestroy(ObjectDestroyMessage msg)
-        {
-            DestroyObject(msg.netId);
-        }
-
-        void DestroyObject(uint netId)
-        {
-            if (logger.LogEnabled()) logger.Log("ClientScene.OnObjDestroy netId:" + netId);
-
-            if (Spawned.TryGetValue(netId, out NetworkIdentity localObject) && localObject != null)
-            {
-                UnSpawn(localObject);
-                Spawned.Remove(netId);
-            }
-            else
-            {
-                logger.LogWarning("Did not find target for destroy message for " + netId);
-            }
-        }
-
-        internal void OnHostClientObjectDestroy(ObjectDestroyMessage msg)
-        {
-            if (logger.LogEnabled()) logger.Log("ClientScene.OnLocalObjectObjDestroy netId:" + msg.netId);
-
-            Spawned.Remove(msg.netId);
-        }
-
-        internal void OnHostClientObjectHide(ObjectHideMessage msg)
-        {
-            if (logger.LogEnabled()) logger.Log("ClientScene::OnLocalObjectObjHide netId:" + msg.netId);
-
-            if (Spawned.TryGetValue(msg.netId, out NetworkIdentity localObject) && localObject != null)
-            {
-                localObject.OnSetHostVisibility(false);
-            }
-        }
-
-        internal void OnHostClientSpawn(SpawnMessage msg)
-        {
-            if (Spawned.TryGetValue(msg.netId, out NetworkIdentity localObject) && localObject != null)
-            {
-                if (msg.isLocalPlayer)
-                    InternalAddPlayer(localObject);
-
-                localObject.HasAuthority = msg.isOwner;
-                localObject.NotifyAuthority();
-                localObject.StartClient();
-                localObject.OnSetHostVisibility(true);
-                CheckForLocalPlayer(localObject);
-            }
-        }
-
-        internal void OnUpdateVarsMessage(UpdateVarsMessage msg)
-        {
-            if (logger.LogEnabled()) logger.Log("ClientScene.OnUpdateVarsMessage " + msg.netId);
-
-            if (Spawned.TryGetValue(msg.netId, out NetworkIdentity localObject) && localObject != null)
-            {
-                using (PooledNetworkReader networkReader = NetworkReaderPool.GetReader(msg.payload))
-                    localObject.OnDeserializeAllSafely(networkReader, false);
-            }
-            else
-            {
-                logger.LogWarning("Did not find target for sync message for " + msg.netId + " . Note: this can be completely normal because UDP messages may arrive out of order, so this message might have arrived after a Destroy message.");
-            }
-        }
-
-        internal void OnRpcMessage(RpcMessage msg)
-        {
-            if (logger.LogEnabled()) logger.Log("ClientScene.OnRPCMessage hash:" + msg.functionHash + " netId:" + msg.netId);
-
-            if (Spawned.TryGetValue(msg.netId, out NetworkIdentity identity) && identity != null)
-            {
-                using (PooledNetworkReader networkReader = NetworkReaderPool.GetReader(msg.payload))
-                    identity.HandleRemoteCall(msg.componentIndex, msg.functionHash, MirrorInvokeType.ClientRpc, networkReader);
-            }
-        }
-
-        void CheckForLocalPlayer(NetworkIdentity identity)
-        {
-            if (identity && identity == LocalPlayer)
-            {
-                // Set isLocalPlayer to true on this NetworkIdentity and trigger OnStartLocalPlayer in all scripts on the same GO
-                identity.ConnectionToServer = Connection;
-                identity.StartLocalPlayer();
-
-                if (logger.LogEnabled()) logger.Log("ClientScene.OnOwnerMessage - player=" + identity.name);
             }
         }
     }
