+++ Assets/Mirror/Runtime/NetworkObjectManager.cs
@@ -1,0 +1,573 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
+using Object = UnityEngine.Object;
+
+namespace Mirror
+{
+
+    [AddComponentMenu("Network/NetworkObjectManager")]
+    [DisallowMultipleComponent]
+    public class NetworkObjectManager : MonoBehaviour
+    {
+        static readonly ILogger logger = LogFactory.GetLogger(typeof(NetworkSceneManager));
+
+        public NetworkClient client;
+        public NetworkServer server;
+        public NetworkSceneManager networkSceneManager;
+
+        public void Start()
+        {
+            if (client != null)
+            {
+                client.Connected.AddListener(OnClientConnected);
+                client.Disconnected.AddListener(OnClientDisconnected);
+
+                if (networkSceneManager != null)
+                    networkSceneManager.ClientSceneChanged.AddListener(OnClientSceneChanged);
+            }
+            if (server != null)
+            {
+
+            }
+        }
+
+        void OnClientConnected(INetworkConnection conn)
+        {
+            RegisterSpawnPrefabs();
+
+            if(client.IsLocalClient)
+            {
+                RegisterHostHandlers();
+            }
+            else
+            {
+                RegisterMessageHandlers();
+            }
+        }
+
+        void OnClientDisconnected()
+        {
+            ClearSpawners();
+            DestroyAllClientObjects();
+        }
+
+        void OnClientSceneChanged(string scenePath, SceneOperation sceneOperation)
+        {
+            PrepareToSpawnSceneObjects();
+        }
+
+        internal void RegisterHostHandlers()
+        {
+            client.Connection.RegisterHandler<ObjectDestroyMessage>(OnHostClientObjectDestroy);
+            client.Connection.RegisterHandler<ObjectHideMessage>(OnHostClientObjectHide);
+            client.Connection.RegisterHandler<SpawnMessage>(OnHostClientSpawn);
+            // host mode reuses objects in the server
+            // so we don't need to spawn them
+            client.Connection.RegisterHandler<UpdateVarsMessage>(msg => { });
+            client.Connection.RegisterHandler<RpcMessage>(OnRpcMessage);
+        }
+
+        internal void RegisterMessageHandlers()
+        {
+            client.Connection.RegisterHandler<ObjectDestroyMessage>(OnObjectDestroy);
+            client.Connection.RegisterHandler<ObjectHideMessage>(OnObjectHide);
+            client.Connection.RegisterHandler<SpawnMessage>(OnSpawn);
+            client.Connection.RegisterHandler<UpdateVarsMessage>(OnUpdateVarsMessage);
+            client.Connection.RegisterHandler<RpcMessage>(OnRpcMessage);
+        }
+
+        static bool ConsiderForSpawning(NetworkIdentity identity)
+        {
+            // not spawned yet, not hidden, etc.?
+            return !identity.gameObject.activeSelf &&
+                   identity.gameObject.hideFlags != HideFlags.NotEditable &&
+                   identity.gameObject.hideFlags != HideFlags.HideAndDontSave &&
+                   identity.sceneId != 0;
+        }
+
+        // this is called from message handler for Owner message
+        internal void InternalAddPlayer(NetworkIdentity identity)
+        {
+            if (client.Connection != null)
+            {
+                client.Connection.Identity = identity;
+            }
+            else
+            {
+                logger.LogWarning("No ready connection found for setting player controller during InternalAddPlayer");
+            }
+        }
+
+        /// <summary>
+        /// Call this after loading/unloading a scene in the client after connection to register the spawnable objects
+        /// </summary>
+        public void PrepareToSpawnSceneObjects()
+        {
+            // add all unspawned NetworkIdentities to spawnable objects
+            client.spawnableObjects.Clear();
+            IEnumerable<NetworkIdentity> sceneObjects =
+                Resources.FindObjectsOfTypeAll<NetworkIdentity>()
+                               .Where(ConsiderForSpawning);
+
+            foreach (NetworkIdentity obj in sceneObjects)
+            {
+                client.spawnableObjects.Add(obj.sceneId, obj);
+            }
+        }
+
+        #region Spawn Prefabs
+        private void RegisterSpawnPrefabs()
+        {
+            for (int i = 0; i < client.spawnPrefabs.Count; i++)
+            {
+                GameObject prefab = client.spawnPrefabs[i];
+                if (prefab != null)
+                {
+                    RegisterPrefab(prefab);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Find the registered prefab for this asset id.
+        /// Useful for debuggers
+        /// </summary>
+        /// <param name="assetId">asset id of the prefab</param>
+        /// <param name="prefab">the prefab gameobject</param>
+        /// <returns>true if prefab was registered</returns>
+        public GameObject GetPrefab(Guid assetId)
+        {
+            if (assetId == Guid.Empty)
+                return null;
+
+            if (client.prefabs.TryGetValue(assetId, out GameObject prefab))
+            {
+                return prefab;
+            }
+            return null;
+        }
+
+        /// <summary>
+        /// Registers a prefab with the spawning system.
+        /// <para>When a NetworkIdentity object is spawned on a server with NetworkServer.SpawnObject(), and the prefab that the object was created from was registered with RegisterPrefab(), the client will use that prefab to instantiate a corresponding client object with the same netId.</para>
+        /// <para>The NetworkManager has a list of spawnable prefabs, it uses this function to register those prefabs with the ClientScene.</para>
+        /// <para>The set of current spawnable object is available in the ClientScene static member variable ClientScene.prefabs, which is a dictionary of NetworkAssetIds and prefab references.</para>
+        /// </summary>
+        /// <param name="prefab">A Prefab that will be spawned.</param>
+        /// <param name="newAssetId">An assetId to be assigned to this prefab. This allows a dynamically created game object to be registered for an already known asset Id.</param>
+        public void RegisterPrefab(GameObject prefab, Guid newAssetId)
+        {
+            NetworkIdentity identity = prefab.GetComponent<NetworkIdentity>();
+            if (identity)
+            {
+                identity.AssetId = newAssetId;
+
+                if (logger.LogEnabled()) logger.Log("Registering prefab '" + prefab.name + "' as asset:" + identity.AssetId);
+                client.prefabs[identity.AssetId] = prefab;
+            }
+            else
+            {
+                throw new InvalidOperationException("Could not register '" + prefab.name + "' since it contains no NetworkIdentity component");
+            }
+        }
+
+        /// <summary>
+        /// Registers a prefab with the spawning system.
+        /// <para>When a NetworkIdentity object is spawned on a server with NetworkServer.SpawnObject(), and the prefab that the object was created from was registered with RegisterPrefab(), the client will use that prefab to instantiate a corresponding client object with the same netId.</para>
+        /// <para>The NetworkManager has a list of spawnable prefabs, it uses this function to register those prefabs with the ClientScene.</para>
+        /// <para>The set of current spawnable object is available in the ClientScene static member variable ClientScene.prefabs, which is a dictionary of NetworkAssetIds and prefab references.</para>
+        /// </summary>
+        /// <param name="prefab">A Prefab that will be spawned.</param>
+        public void RegisterPrefab(GameObject prefab)
+        {
+            NetworkIdentity identity = prefab.GetComponent<NetworkIdentity>();
+            if (identity)
+            {
+                if (logger.LogEnabled()) logger.Log("Registering prefab '" + prefab.name + "' as asset:" + identity.AssetId);
+                client.prefabs[identity.AssetId] = prefab;
+
+                NetworkIdentity[] identities = prefab.GetComponentsInChildren<NetworkIdentity>();
+                if (identities.Length > 1)
+                {
+                    logger.LogWarning("The prefab '" + prefab.name +
+                                     "' has multiple NetworkIdentity components. There can only be one NetworkIdentity on a prefab, and it must be on the root object.");
+                }
+            }
+            else
+            {
+                throw new InvalidOperationException("Could not register '" + prefab.name + "' since it contains no NetworkIdentity component");
+            }
+        }
+
+        /// <summary>
+        /// Registers a prefab with the spawning system.
+        /// <para>When a NetworkIdentity object is spawned on a server with NetworkServer.SpawnObject(), and the prefab that the object was created from was registered with RegisterPrefab(), the client will use that prefab to instantiate a corresponding client object with the same netId.</para>
+        /// <para>The NetworkManager has a list of spawnable prefabs, it uses this function to register those prefabs with the ClientScene.</para>
+        /// <para>The set of current spawnable object is available in the ClientScene static member variable ClientScene.prefabs, which is a dictionary of NetworkAssetIds and prefab references.</para>
+        /// </summary>
+        /// <param name="prefab">A Prefab that will be spawned.</param>
+        /// <param name="spawnHandler">A method to use as a custom spawnhandler on clients.</param>
+        /// <param name="unspawnHandler">A method to use as a custom un-spawnhandler on clients.</param>
+        public void RegisterPrefab(GameObject prefab, SpawnDelegate spawnHandler, UnSpawnDelegate unspawnHandler)
+        {
+            RegisterPrefab(prefab, msg => spawnHandler(msg.position, msg.assetId), unspawnHandler);
+        }
+
+        /// <summary>
+        /// Registers a prefab with the spawning system.
+        /// <para>When a NetworkIdentity object is spawned on a server with NetworkServer.SpawnObject(), and the prefab that the object was created from was registered with RegisterPrefab(), the client will use that prefab to instantiate a corresponding client object with the same netId.</para>
+        /// <para>The NetworkManager has a list of spawnable prefabs, it uses this function to register those prefabs with the ClientScene.</para>
+        /// <para>The set of current spawnable object is available in the ClientScene static member variable ClientScene.prefabs, which is a dictionary of NetworkAssetIds and prefab references.</para>
+        /// </summary>
+        /// <param name="prefab">A Prefab that will be spawned.</param>
+        /// <param name="spawnHandler">A method to use as a custom spawnhandler on clients.</param>
+        /// <param name="unspawnHandler">A method to use as a custom un-spawnhandler on clients.</param>
+        public void RegisterPrefab(GameObject prefab, SpawnHandlerDelegate spawnHandler, UnSpawnDelegate unspawnHandler)
+        {
+            NetworkIdentity identity = prefab.GetComponent<NetworkIdentity>();
+            if (identity == null)
+            {
+                throw new InvalidOperationException("Could not register '" + prefab.name + "' since it contains no NetworkIdentity component");
+            }
+
+            if (identity.AssetId == Guid.Empty)
+            {
+                throw new InvalidOperationException("RegisterPrefab game object " + prefab.name + " has no " + nameof(prefab) + ". Use RegisterSpawnHandler() instead?");
+            }
+
+            if (logger.LogEnabled()) logger.Log("Registering custom prefab '" + prefab.name + "' as asset:" + identity.AssetId + " " + spawnHandler.GetMethodName() + "/" + unspawnHandler.GetMethodName());
+
+            client.spawnHandlers[identity.AssetId] = spawnHandler;
+            client.unspawnHandlers[identity.AssetId] = unspawnHandler;
+        }
+
+        /// <summary>
+        /// Removes a registered spawn prefab that was setup with ClientScene.RegisterPrefab.
+        /// </summary>
+        /// <param name="prefab">The prefab to be removed from registration.</param>
+        public void UnregisterPrefab(GameObject prefab)
+        {
+            NetworkIdentity identity = prefab.GetComponent<NetworkIdentity>();
+            if (identity == null)
+            {
+                throw new InvalidOperationException("Could not unregister '" + prefab.name + "' since it contains no NetworkIdentity component");
+            }
+            client.spawnHandlers.Remove(identity.AssetId);
+            client.unspawnHandlers.Remove(identity.AssetId);
+        }
+
+        #endregion
+
+        #region Spawn Handler
+
+        /// <summary>
+        /// This is an advanced spawning function that registers a custom assetId with the UNET spawning system.
+        /// <para>This can be used to register custom spawning methods for an assetId - instead of the usual method of registering spawning methods for a prefab. This should be used when no prefab exists for the spawned objects - such as when they are constructed dynamically at runtime from configuration data.</para>
+        /// </summary>
+        /// <param name="assetId">Custom assetId string.</param>
+        /// <param name="spawnHandler">A method to use as a custom spawnhandler on clients.</param>
+        /// <param name="unspawnHandler">A method to use as a custom un-spawnhandler on clients.</param>
+        public void RegisterSpawnHandler(Guid assetId, SpawnDelegate spawnHandler, UnSpawnDelegate unspawnHandler)
+        {
+            RegisterSpawnHandler(assetId, msg => spawnHandler(msg.position, msg.assetId), unspawnHandler);
+        }
+
+        /// <summary>
+        /// This is an advanced spawning function that registers a custom assetId with the UNET spawning system.
+        /// <para>This can be used to register custom spawning methods for an assetId - instead of the usual method of registering spawning methods for a prefab. This should be used when no prefab exists for the spawned objects - such as when they are constructed dynamically at runtime from configuration data.</para>
+        /// </summary>
+        /// <param name="assetId">Custom assetId string.</param>
+        /// <param name="spawnHandler">A method to use as a custom spawnhandler on clients.</param>
+        /// <param name="unspawnHandler">A method to use as a custom un-spawnhandler on clients.</param>
+        public void RegisterSpawnHandler(Guid assetId, SpawnHandlerDelegate spawnHandler, UnSpawnDelegate unspawnHandler)
+        {
+            if (logger.LogEnabled()) logger.Log("RegisterSpawnHandler asset '" + assetId + "' " + spawnHandler.GetMethodName() + "/" + unspawnHandler.GetMethodName());
+
+            client.spawnHandlers[assetId] = spawnHandler;
+            client.unspawnHandlers[assetId] = unspawnHandler;
+        }
+
+        /// <summary>
+        /// Removes a registered spawn handler function that was registered with ClientScene.RegisterHandler().
+        /// </summary>
+        /// <param name="assetId">The assetId for the handler to be removed for.</param>
+        public void UnregisterSpawnHandler(Guid assetId)
+        {
+            client.spawnHandlers.Remove(assetId);
+            client.unspawnHandlers.Remove(assetId);
+        }
+
+        /// <summary>
+        /// This clears the registered spawn prefabs and spawn handler functions for this client.
+        /// </summary>
+        public void ClearSpawners()
+        {
+            client.prefabs.Clear();
+            client.spawnHandlers.Clear();
+            client.unspawnHandlers.Clear();
+        }
+
+        #endregion
+
+        void UnSpawn(NetworkIdentity identity)
+        {
+            Guid assetId = identity.AssetId;
+
+            identity.StopClient();
+            if (client.unspawnHandlers.TryGetValue(assetId, out UnSpawnDelegate handler) && handler != null)
+            {
+                handler(identity.gameObject);
+            }
+            else if (identity.sceneId == 0)
+            {
+                Destroy(identity.gameObject);
+            }
+            else
+            {
+                identity.Reset();
+                identity.gameObject.SetActive(false);
+                client.spawnableObjects[identity.sceneId] = identity;
+            }
+        }
+
+        /// <summary>
+        /// Destroys all networked objects on the client.
+        /// <para>This can be used to clean up when a network connection is closed.</para>
+        /// </summary>
+        public void DestroyAllClientObjects()
+        {
+            foreach (NetworkIdentity identity in client.Spawned.Values)
+            {
+                if (identity != null && identity.gameObject != null)
+                {
+                    UnSpawn(identity);
+                }
+            }
+            client.Spawned.Clear();
+        }
+
+        void ApplySpawnPayload(NetworkIdentity identity, SpawnMessage msg)
+        {
+            if (msg.assetId != Guid.Empty)
+                identity.AssetId = msg.assetId;
+
+            if (!identity.gameObject.activeSelf)
+            {
+                identity.gameObject.SetActive(true);
+            }
+
+            // apply local values for VR support
+            identity.transform.localPosition = msg.position;
+            identity.transform.localRotation = msg.rotation;
+            identity.transform.localScale = msg.scale;
+            identity.HasAuthority = msg.isOwner;
+            identity.NetId = msg.netId;
+            identity.Client = client;
+
+            if (msg.isLocalPlayer)
+                InternalAddPlayer(identity);
+
+            // deserialize components if any payload
+            // (Count is 0 if there were no components)
+            if (msg.payload.Count > 0)
+            {
+                using (PooledNetworkReader payloadReader = NetworkReaderPool.GetReader(msg.payload))
+                {
+                    identity.OnDeserializeAllSafely(payloadReader, true);
+                }
+            }
+
+            client.Spawned[msg.netId] = identity;
+
+            // objects spawned as part of initial state are started on a second pass
+            identity.NotifyAuthority();
+            identity.StartClient();
+            CheckForLocalPlayer(identity);
+        }
+
+        internal void OnSpawn(SpawnMessage msg)
+        {
+            if (msg.assetId == Guid.Empty && msg.sceneId == 0)
+            {
+                throw new InvalidOperationException("OnObjSpawn netId: " + msg.netId + " has invalid asset Id");
+            }
+            if (logger.LogEnabled()) logger.Log($"Client spawn handler instantiating netId={msg.netId} assetID={msg.assetId} sceneId={msg.sceneId} pos={msg.position}");
+
+            // was the object already spawned?
+            NetworkIdentity identity = GetExistingObject(msg.netId);
+
+            if (identity == null)
+            {
+                identity = msg.sceneId == 0 ? SpawnPrefab(msg) : SpawnSceneObject(msg);
+            }
+
+            if (identity == null)
+            {
+                throw new InvalidOperationException($"Could not spawn assetId={msg.assetId} scene={msg.sceneId} netId={msg.netId}");
+            }
+
+            ApplySpawnPayload(identity, msg);
+        }
+
+        NetworkIdentity GetExistingObject(uint netid)
+        {
+            client.Spawned.TryGetValue(netid, out NetworkIdentity localObject);
+            return localObject;
+        }
+
+        NetworkIdentity SpawnPrefab(SpawnMessage msg)
+        {
+            if (client.spawnHandlers.TryGetValue(msg.assetId, out SpawnHandlerDelegate handler))
+            {
+                GameObject obj = handler(msg);
+                if (obj == null)
+                {
+                    logger.LogWarning("Client spawn handler for " + msg.assetId + " returned null");
+                    return null;
+                }
+                return obj.GetComponent<NetworkIdentity>();
+            }
+            GameObject prefab = GetPrefab(msg.assetId);
+            if ((object)prefab != null)
+            {
+                GameObject obj = Object.Instantiate(prefab, msg.position, msg.rotation);
+                if (logger.LogEnabled())
+                {
+                    logger.Log("Client spawn handler instantiating [netId:" + msg.netId + " asset ID:" + msg.assetId + " pos:" + msg.position + " rotation: " + msg.rotation + "]");
+                }
+
+                return obj.GetComponent<NetworkIdentity>();
+            }
+            logger.LogError("Failed to spawn server object, did you forget to add it to the NetworkManager? assetId=" + msg.assetId + " netId=" + msg.netId);
+            return null;
+        }
+
+        NetworkIdentity SpawnSceneObject(SpawnMessage msg)
+        {
+            NetworkIdentity spawnedId = SpawnSceneObject(msg.sceneId);
+            if (spawnedId == null)
+            {
+                logger.LogError("Spawn scene object not found for " + msg.sceneId.ToString("X") + " SpawnableObjects.Count=" + client.spawnableObjects.Count);
+
+                // dump the whole spawnable objects dict for easier debugging
+                if (logger.LogEnabled())
+                {
+                    foreach (KeyValuePair<ulong, NetworkIdentity> kvp in client.spawnableObjects)
+                        logger.Log("Spawnable: SceneId=" + kvp.Key + " name=" + kvp.Value.name);
+                }
+            }
+
+            if (logger.LogEnabled()) logger.Log("Client spawn for [netId:" + msg.netId + "] [sceneId:" + msg.sceneId + "] obj:" + spawnedId);
+            return spawnedId;
+        }
+
+        NetworkIdentity SpawnSceneObject(ulong sceneId)
+        {
+            if (client.spawnableObjects.TryGetValue(sceneId, out NetworkIdentity identity))
+            {
+                client.spawnableObjects.Remove(sceneId);
+                return identity;
+            }
+            logger.LogWarning("Could not find scene object with sceneid:" + sceneId.ToString("X"));
+            return null;
+        }
+
+        internal void OnObjectHide(ObjectHideMessage msg)
+        {
+            DestroyObject(msg.netId);
+        }
+
+        internal void OnObjectDestroy(ObjectDestroyMessage msg)
+        {
+            DestroyObject(msg.netId);
+        }
+
+        void DestroyObject(uint netId)
+        {
+            if (logger.LogEnabled()) logger.Log("ClientScene.OnObjDestroy netId:" + netId);
+
+            if (client.Spawned.TryGetValue(netId, out NetworkIdentity localObject) && localObject != null)
+            {
+                UnSpawn(localObject);
+                client.Spawned.Remove(netId);
+            }
+            else
+            {
+                logger.LogWarning("Did not find target for destroy message for " + netId);
+            }
+        }
+
+        internal void OnHostClientObjectDestroy(ObjectDestroyMessage msg)
+        {
+            if (logger.LogEnabled()) logger.Log("ClientScene.OnLocalObjectObjDestroy netId:" + msg.netId);
+
+            client.Spawned.Remove(msg.netId);
+        }
+
+        internal void OnHostClientObjectHide(ObjectHideMessage msg)
+        {
+            if (logger.LogEnabled()) logger.Log("ClientScene::OnLocalObjectObjHide netId:" + msg.netId);
+
+            if (client.Spawned.TryGetValue(msg.netId, out NetworkIdentity localObject) && localObject != null)
+            {
+                localObject.OnSetHostVisibility(false);
+            }
+        }
+
+        internal void OnHostClientSpawn(SpawnMessage msg)
+        {
+            if (client.Spawned.TryGetValue(msg.netId, out NetworkIdentity localObject) && localObject != null)
+            {
+                if (msg.isLocalPlayer)
+                    InternalAddPlayer(localObject);
+
+                localObject.HasAuthority = msg.isOwner;
+                localObject.NotifyAuthority();
+                localObject.StartClient();
+                localObject.OnSetHostVisibility(true);
+                CheckForLocalPlayer(localObject);
+            }
+        }
+
+        internal void OnUpdateVarsMessage(UpdateVarsMessage msg)
+        {
+            if (logger.LogEnabled()) logger.Log("ClientScene.OnUpdateVarsMessage " + msg.netId);
+
+            if (client.Spawned.TryGetValue(msg.netId, out NetworkIdentity localObject) && localObject != null)
+            {
+                using (PooledNetworkReader networkReader = NetworkReaderPool.GetReader(msg.payload))
+                    localObject.OnDeserializeAllSafely(networkReader, false);
+            }
+            else
+            {
+                logger.LogWarning("Did not find target for sync message for " + msg.netId + " . Note: this can be completely normal because UDP messages may arrive out of order, so this message might have arrived after a Destroy message.");
+            }
+        }
+
+        internal void OnRpcMessage(RpcMessage msg)
+        {
+            if (logger.LogEnabled()) logger.Log("ClientScene.OnRPCMessage hash:" + msg.functionHash + " netId:" + msg.netId);
+
+            if (client.Spawned.TryGetValue(msg.netId, out NetworkIdentity identity) && identity != null)
+            {
+                using (PooledNetworkReader networkReader = NetworkReaderPool.GetReader(msg.payload))
+                    identity.HandleRemoteCall(msg.componentIndex, msg.functionHash, MirrorInvokeType.ClientRpc, networkReader);
+            }
+        }
+
+        void CheckForLocalPlayer(NetworkIdentity identity)
+        {
+            if (identity && identity == client.LocalPlayer)
+            {
+                // Set isLocalPlayer to true on this NetworkIdentity and trigger OnStartLocalPlayer in all scripts on the same GO
+                identity.ConnectionToServer = client.Connection;
+                identity.StartLocalPlayer();
+
+                if (logger.LogEnabled()) logger.Log("ClientScene.OnOwnerMessage - player=" + identity.name);
+            }
+        }
+    }
+}
+
